new revert

+++ ear/src/main/java/cz/cvut/fel/ear/dao/ReviewRepository.java
@@ -8,7 +8,5 @@ import java.util.List;
 
 @Repository
 public interface ReviewRepository extends JpaRepository<Review, Long> {
-    // TODO - add comment
-    List<Review> findAllByBoardGame_IdIn(List<Long> integers);
-
+    List<Review> findAllByBoardGame_IdIn(List<Integer> integers);
 }
